requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
line-bot-sdk==3.9.0
firebase-admin==6.4.0
httpx==0.26.0
stripe==7.14.0 